Implemented linking list of lines

+++ src/main/java/it/polito/ai/es02/model/BusLineStop.java
@@ -37,7 +37,7 @@ public class BusLineStop implements Comparable<BusLineStop>{
 	public void setBusStop(BusStop busStop){
 		this.primaryKey.setBusStop(busStop);
 	}
-	public int getSequenceNumber() {
+	public short getSequenceNumber() {
 		return this.primaryKey.getSequenceNumber();
 	}
 	public void setSequenceNumber(short sequenceNumber) {
